Move Block and related types to separate file
